fix recording CartItem twice
1. When a user is authenticated and adds a cart item (POST /api/cart),
it is added to the database.
2. Then the same user adds the item as an unauthenticated user, the item
is added to cookies for the request session.
3. As the user authenticates and checks out the cart (/billing/cart/),
the request session items are indavertently added to the database because
email == '' in the request session and email is None in the database.

This commit fixes the issue by improving the filtering algorithm merging
request session items and database items.